# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Board Game Factory API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List all rules for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
    post:
      operationId: createRule
      tags: [rules]
      summary: Create a new rule for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleBody"
      responses:
        "201":
          description: Created rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
  /projects/{projectId}/rules/{id}:
    patch:
      operationId: updateRule
      tags: [rules]
      summary: Update a rule
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRuleBody"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRule
      tags: [rules]
      summary: Delete a rule
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/rules-sandbox:
    post:
      operationId: sendRulesSandboxMessage
      tags: [rules]
      summary: Send a message to the rules sandbox AI (SSE stream with full project context)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RulesSandboxMessageBody"
      responses:
        "200":
          description: SSE stream of AI response chunks
          content:
            text/event-stream: {}
  /projects/{projectId}/rules-sandbox/history:
    get:
      operationId: getRulesSandboxHistory
      tags: [rules]
      summary: Get rules sandbox chat history for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SandboxMessage"
    delete:
      operationId: clearRulesSandboxHistory
      tags: [rules]
      summary: Clear rules sandbox chat history
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cleared

  # ── Simulation ────────────────────────────────────────────────────────────
  /projects/{projectId}/simulate:
    post:
      operationId: runSimulation
      tags: [simulation]
      summary: Run a Monte Carlo economic simulation
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationConfig"
      responses:
        "200":
          description: Simulation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"
  /projects/{projectId}/simulation-history:
    get:
      operationId: listSimulations
      tags: [simulation]
      summary: List saved simulation runs
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of simulation runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimulationRun"

  # ── Assets ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Projects ──────────────────────────────────────────────────────────
...
    Rule:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: ["string", "null"]
          description: "e.g. movement, combat, economy, turn_structure"
        priority:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - content
        - priority
        - createdAt
        - updatedAt
    CreateRuleBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: ["string", "null"]
        priority:
          type: integer
      required:
        - title
        - content
    UpdateRuleBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        priority:
          type: ["integer", "null"]
    RulesSandboxMessageBody:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    SandboxMessage:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - role
        - content
        - createdAt

    # ── Simulation ────────────────────────────────────────────────────────
    SimulationConfig:
      type: object
      properties:
        iterations:
          type: integer
          description: Number of Monte Carlo iterations (default 1000)
        turns:
          type: integer
          description: Number of game turns per simulation
        goldPerTurn:
          type: number
          description: Base gold income per turn
        goldSpendPerTurn:
          type: number
          description: Base gold spend per turn
        incomeVariance:
          type: number
          description: Variance factor for income (0-1)
        spendVariance:
          type: number
          description: Variance factor for spending (0-1)
        startingGold:
          type: number
        label:
          type: ["string", "null"]
      required:
        - iterations
        - turns
        - goldPerTurn
        - goldSpendPerTurn
    SimulationResult:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        label:
          type: ["string", "null"]
        config:
          $ref: "#/components/schemas/SimulationConfig"
        percentile10:
          type: number
        percentile50:
          type: number
        percentile90:
          type: number
        mean:
          type: number
        stdDev:
          type: number
        economyHealthScore:
          type: number
          description: "0-100 score: >70 balanced, 40-70 unbalanced, <40 broken"
        turnData:
          type: array
          items:
            $ref: "#/components/schemas/TurnDataPoint"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - config
        - percentile10
        - percentile50
        - percentile90
        - mean
        - stdDev
        - economyHealthScore
        - turnData
        - createdAt
    TurnDataPoint:
      type: object
      properties:
        turn:
          type: integer
        p10:
          type: number
        p50:
          type: number
        p90:
          type: number
        mean:
          type: number
      required:
        - turn
        - p10
        - p50
        - p90
        - mean
    SimulationRun:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        label:
          type: ["string", "null"]
        economyHealthScore:
          type: number
        mean:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - economyHealthScore
        - mean
        - createdAt

    # ── Assets ────────────────────────────────────────────────────────────
...